refactor(settings): split video settings responsibilities
