+++ src/DownKyi.Desktop/ViewModels/Settings/ViewVideoViewModel.cs
@@ -1,9 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Collections.ObjectModel;
 using System.Linq;
-using System.Threading.Tasks;
 using CommunityToolkit.Mvvm.Input;
 using DownKyi.Application.Desktop;
 using DownKyi.Commands;
@@ -22,7 +20,7 @@ internal sealed class FfmpegHardwareAccelerationItem
     public FfmpegHardwareAcceleration Value { get; init; }
 }
 
-internal class ViewVideoViewModel : ViewModelBase
+internal partial class ViewVideoViewModel : ViewModelBase
 {
     public const string Tag = "PageSettingsVideo";
 
@@ -30,243 +28,6 @@ internal class ViewVideoViewModel : ViewModelBase
     private readonly IFilePickerService _filePickerService;
     private readonly ILogger<ViewVideoViewModel> _logger;
     private readonly ISettingsStore _settingsStore;
-
-    #region 页面属性申明
-
-    private IReadOnlyList<Quality> _videoCodecs = Array.Empty<Quality>();
-
-    public IReadOnlyList<Quality> VideoCodecs
-    {
-        get => _videoCodecs;
-        set => SetProperty(ref _videoCodecs, value);
-    }
-
-    private Quality _selectedVideoCodec = null!;
-
-    public Quality SelectedVideoCodec
-    {
-        get => _selectedVideoCodec;
-        set => SetProperty(ref _selectedVideoCodec, value);
-    }
-
-    private IReadOnlyList<Quality> _videoQualityList = Array.Empty<Quality>();
-
-    public IReadOnlyList<Quality> VideoQualityList
-    {
-        get => _videoQualityList;
-        set => SetProperty(ref _videoQualityList, value);
-    }
-
-    private Quality _selectedVideoQuality = null!;
-
-    public Quality SelectedVideoQuality
-    {
-        get => _selectedVideoQuality;
-        set => SetProperty(ref _selectedVideoQuality, value);
-    }
-
-    private IReadOnlyList<Quality> _audioQualityList = Array.Empty<Quality>();
-
-    public IReadOnlyList<Quality> AudioQualityList
-    {
-        get => _audioQualityList;
-        set => SetProperty(ref _audioQualityList, value);
-    }
-
-    private Quality _selectedAudioQuality = null!;
-
-    public Quality SelectedAudioQuality
-    {
-        get => _selectedAudioQuality;
-        set => SetProperty(ref _selectedAudioQuality, value);
-    }
-
-    private IReadOnlyList<VideoParseType> _videoParseTypeList = Array.Empty<VideoParseType>();
-
-    public IReadOnlyList<VideoParseType> VideoParseTypeList
-    {
-        get => _videoParseTypeList;
-        set => SetProperty(ref _videoParseTypeList, value);
-    }
-
-    private VideoParseType _selectedVideoParseType = null!;
-
-    public VideoParseType SelectedVideoParseType
-    {
-        get => _selectedVideoParseType;
-        set => SetProperty(ref _selectedVideoParseType, value);
-    }
-
-    private bool _isTranscodingFlvToMp4;
-
-    public bool IsTranscodingFlvToMp4
-    {
-        get => _isTranscodingFlvToMp4;
-        set => SetProperty(ref _isTranscodingFlvToMp4, value);
-    }
-
-    private bool _isTranscodingAacToMp3;
-
-    public bool IsTranscodingAacToMp3
-    {
-        get => _isTranscodingAacToMp3;
-        set => SetProperty(ref _isTranscodingAacToMp3, value);
-    }
-
-    private IReadOnlyList<FfmpegHardwareAccelerationItem> _ffmpegHardwareAccelerations = Array.Empty<FfmpegHardwareAccelerationItem>();
-
-    public IReadOnlyList<FfmpegHardwareAccelerationItem> FfmpegHardwareAccelerations
-    {
-        get => _ffmpegHardwareAccelerations;
-        set => SetProperty(ref _ffmpegHardwareAccelerations, value);
-    }
-
-    private FfmpegHardwareAccelerationItem _selectedFfmpegHardwareAcceleration = null!;
-
-    public FfmpegHardwareAccelerationItem SelectedFfmpegHardwareAcceleration
-    {
-        get => _selectedFfmpegHardwareAcceleration;
-        set => SetProperty(ref _selectedFfmpegHardwareAcceleration, value);
-    }
-
-    private IReadOnlyList<int> _ffmpegMaxParallelJobs = Array.Empty<int>();
-
-    public IReadOnlyList<int> FfmpegMaxParallelJobs
-    {
-        get => _ffmpegMaxParallelJobs;
-        set => SetProperty(ref _ffmpegMaxParallelJobs, value);
-    }
-
-    private int _selectedFfmpegMaxParallelJob;
-
-    public int SelectedFfmpegMaxParallelJob
-    {
-        get => _selectedFfmpegMaxParallelJob;
-        set => SetProperty(ref _selectedFfmpegMaxParallelJob, value);
-    }
-
-    private bool _isUseDefaultDirectory;
-
-    public bool IsUseDefaultDirectory
-    {
-        get => _isUseDefaultDirectory;
-        set => SetProperty(ref _isUseDefaultDirectory, value);
-    }
-
-    private string _saveVideoDirectory = string.Empty;
-
-    public string SaveVideoDirectory
-    {
-        get => _saveVideoDirectory;
-        set => SetProperty(ref _saveVideoDirectory, value);
-    }
-
-    private bool _downloadAll;
-
-    public bool DownloadAll
-    {
-        get => _downloadAll;
-        set => SetProperty(ref _downloadAll, value);
-    }
-
-    private bool _downloadAudio;
-
-    public bool DownloadAudio
-    {
-        get => _downloadAudio;
-        set => SetProperty(ref _downloadAudio, value);
-    }
-
-    private bool _downloadVideo;
-
-    public bool DownloadVideo
-    {
-        get => _downloadVideo;
-        set => SetProperty(ref _downloadVideo, value);
-    }
-
-    private bool _downloadDanmaku;
-
-    public bool DownloadDanmaku
-    {
-        get => _downloadDanmaku;
-        set => SetProperty(ref _downloadDanmaku, value);
-    }
-
-    private bool _downloadSubtitle;
-
-    public bool DownloadSubtitle
-    {
-        get => _downloadSubtitle;
-        set => SetProperty(ref _downloadSubtitle, value);
-    }
-
-    private bool _downloadCover;
-
-    public bool DownloadCover
-    {
-        get => _downloadCover;
-        set => SetProperty(ref _downloadCover, value);
-    }
-
-
-    private bool _generateMovieMetadata;
-
-    public bool GenerateMovieMetadata
-    {
-        get => _generateMovieMetadata;
-        set => SetProperty(ref _generateMovieMetadata, value);
-    }
-
-    private ObservableCollection<DisplayFileNamePart> _selectedFileName = new();
-
-    public ObservableCollection<DisplayFileNamePart> SelectedFileName => _selectedFileName;
-
-    private ObservableCollection<DisplayFileNamePart> _optionalFields = new();
-
-    public ObservableCollection<DisplayFileNamePart> OptionalFields => _optionalFields;
-
-    private int _selectedOptionalField;
-
-    public int SelectedOptionalField
-    {
-        get => _selectedOptionalField;
-        set => SetProperty(ref _selectedOptionalField, value);
-    }
-
-    private IReadOnlyList<string> _fileNamePartTimeFormatList = Array.Empty<string>();
-
-    public IReadOnlyList<string> FileNamePartTimeFormatList
-    {
-        get => _fileNamePartTimeFormatList;
-        set => SetProperty(ref _fileNamePartTimeFormatList, value);
-    }
-
-    private string _selectedFileNamePartTimeFormat = string.Empty;
-
-    public string SelectedFileNamePartTimeFormat
-    {
-        get => _selectedFileNamePartTimeFormat;
-        set => SetProperty(ref _selectedFileNamePartTimeFormat, value);
-    }
-
-    private IReadOnlyList<OrderFormatDisplay> _orderFormatList = Array.Empty<OrderFormatDisplay>();
-
-    public IReadOnlyList<OrderFormatDisplay> OrderFormatList
-    {
-        get => _orderFormatList;
-        set => SetProperty(ref _orderFormatList, value);
-    }
-
-    private OrderFormatDisplay _orderFormatDisplay = null!;
-
-    public OrderFormatDisplay OrderFormatDisplay
-    {
-        get => _orderFormatDisplay;
-        set => SetProperty(ref _orderFormatDisplay, value);
-    }
-
-    #endregion
 
     public ViewVideoViewModel(
         IDesktopInteractionContext desktopInteractions,
@@ -594,336 +355,6 @@ internal class ViewVideoViewModel : ViewModelBase
         {
             FfmpegMaxParallelJobs = maxParallelJobs
         }).FfmpegMaxParallelJobs == maxParallelJobs;
-        PublishTip(isSucceed);
-    }
-
-    // 是否使用默认下载目录事件
-    private RelayCommand? _isUseDefaultDirectoryCommand;
-
-    public RelayCommand IsUseDefaultDirectoryCommand => _isUseDefaultDirectoryCommand ??= new RelayCommand(ExecuteIsUseDefaultDirectoryCommand);
-
-    /// <summary>
-    /// 是否使用默认下载目录事件
-    /// </summary>
-    private void ExecuteIsUseDefaultDirectoryCommand()
-    {
-        var isUseDefaultDirectory = IsUseDefaultDirectory ? AllowStatus.Yes : AllowStatus.No;
-
-        var isSucceed = UpdateVideo(settings => settings with
-        {
-            IsUseSaveVideoRootPath = isUseDefaultDirectory
-        }).IsUseSaveVideoRootPath == isUseDefaultDirectory;
-        PublishTip(isSucceed);
-    }
-
-    // 修改默认下载目录事件
-    private DownKyiAsyncDelegateCommand? _changeSaveVideoDirectoryCommand;
-
-    public DownKyiAsyncDelegateCommand ChangeSaveVideoDirectoryCommand => _changeSaveVideoDirectoryCommand ??= new DownKyiAsyncDelegateCommand(ExecuteChangeSaveVideoDirectoryCommand, _logger);
-
-    /// <summary>
-    /// 修改默认下载目录事件
-    /// </summary>
-    private async Task ExecuteChangeSaveVideoDirectoryCommand()
-    {
-        var directory = await _filePickerService.SelectFolderAsync().ConfigureAwait(true);
-        if (string.IsNullOrEmpty(directory))
-        {
-            return;
-        }
-
-        var isSucceed = UpdateVideo(settings => settings with
-        {
-            SaveVideoRootPath = directory
-        }).SaveVideoRootPath == directory;
-        PublishTip(isSucceed);
-
-        if (isSucceed)
-        {
-            SaveVideoDirectory = directory;
-        }
-    }
-
-    // 所有内容选择事件
-    private RelayCommand? _downloadAllCommand;
-
-    public RelayCommand DownloadAllCommand => _downloadAllCommand ??= new RelayCommand(ExecuteDownloadAllCommand);
-
-    /// <summary>
-    /// 所有内容选择事件
-    /// </summary>
-    private void ExecuteDownloadAllCommand()
-    {
-        if (DownloadAll)
-        {
-            DownloadAudio = true;
-            DownloadVideo = true;
-            DownloadDanmaku = true;
-            DownloadSubtitle = true;
-            DownloadCover = true;
-        }
-        else
-        {
-            DownloadAudio = false;
-            DownloadVideo = false;
-            DownloadDanmaku = false;
-            DownloadSubtitle = false;
-            DownloadCover = false;
-        }
-
-        SetVideoContent();
-    }
-
-    // 音频选择事件
-    private RelayCommand? _downloadAudioCommand;
-
-    public RelayCommand DownloadAudioCommand => _downloadAudioCommand ??= new RelayCommand(ExecuteDownloadAudioCommand);
-
-    /// <summary>
-    /// 音频选择事件
-    /// </summary>
-    private void ExecuteDownloadAudioCommand()
-    {
-        if (!DownloadAudio)
-        {
-            DownloadAll = false;
-        }
-
-        if (DownloadAudio && DownloadVideo && DownloadDanmaku && DownloadSubtitle && DownloadCover)
-        {
-            DownloadAll = true;
-        }
-
-        SetVideoContent();
-    }
-
-    // 视频选择事件
-    private RelayCommand? _downloadVideoCommand;
-
-    public RelayCommand DownloadVideoCommand => _downloadVideoCommand ??= new RelayCommand(ExecuteDownloadVideoCommand);
-
-    /// <summary>
-    /// 视频选择事件
-    /// </summary>
-    private void ExecuteDownloadVideoCommand()
-    {
-        if (!DownloadVideo)
-        {
-            DownloadAll = false;
-        }
-
-        if (DownloadAudio && DownloadVideo && DownloadDanmaku && DownloadSubtitle && DownloadCover)
-        {
-            DownloadAll = true;
-        }
-
-        SetVideoContent();
-    }
-
-    // 弹幕选择事件
-    private RelayCommand? _downloadDanmakuCommand;
-
-    public RelayCommand DownloadDanmakuCommand => _downloadDanmakuCommand ??= new RelayCommand(ExecuteDownloadDanmakuCommand);
-
-    /// <summary>
-    /// 弹幕选择事件
-    /// </summary>
-    private void ExecuteDownloadDanmakuCommand()
-    {
-        if (!DownloadDanmaku)
-        {
-            DownloadAll = false;
-        }
-
-        if (DownloadAudio && DownloadVideo && DownloadDanmaku && DownloadSubtitle && DownloadCover)
-        {
-            DownloadAll = true;
-        }
-
-        SetVideoContent();
-    }
-
-    // 字幕选择事件
-    private RelayCommand? _downloadSubtitleCommand;
-
-    public RelayCommand DownloadSubtitleCommand => _downloadSubtitleCommand ??= new RelayCommand(ExecuteDownloadSubtitleCommand);
-
-    /// <summary>
-    /// 字幕选择事件
-    /// </summary>
-    private void ExecuteDownloadSubtitleCommand()
-    {
-        if (!DownloadSubtitle)
-        {
-            DownloadAll = false;
-        }
-
-        if (DownloadAudio && DownloadVideo && DownloadDanmaku && DownloadSubtitle && DownloadCover)
-        {
-            DownloadAll = true;
-        }
-
-        SetVideoContent();
-    }
-
-
-    // 封面选择事件
-    private RelayCommand? _downloadCoverCommand;
-
-    public RelayCommand DownloadCoverCommand => _downloadCoverCommand ??= new RelayCommand(ExecuteDownloadCoverCommand);
-
-
-    private RelayCommand? _generateMovieMetadataCommand;
-
-    public RelayCommand GenerateMovieMetadataCommand => _generateMovieMetadataCommand ??= new RelayCommand(ExecuteGenerateMovieMetadataCommand);
-
-
-    private void ExecuteGenerateMovieMetadataCommand()
-    {
-        SetVideoContent();
-    }
-    /// <summary>
-    /// 封面选择事件
-    /// </summary>
-    private void ExecuteDownloadCoverCommand()
-    {
-        if (!DownloadCover)
-        {
-            DownloadAll = false;
-        }
-
-        if (DownloadAudio && DownloadVideo && DownloadDanmaku && DownloadSubtitle && DownloadCover)
-        {
-            DownloadAll = true;
-        }
-
-        SetVideoContent();
-    }
-
-    // 选中文件名字段右键点击事件
-    private RelayCommand<object>? _selectedFileNameRightCommand;
-
-    public RelayCommand<object> SelectedFileNameRightCommand => _selectedFileNameRightCommand ??= RequiredParameterCommand.Create<object>(ExecuteSelectedFileNameRightCommand);
-
-    /// <summary>
-    /// 选中文件名字段右键点击事件
-    /// </summary>
-    /// <param name="parameter"></param>
-    private void ExecuteSelectedFileNameRightCommand(object parameter)
-    {
-        if (parameter == null)
-        {
-            return;
-        }
-
-        var isSucceed = SelectedFileName.Remove((DisplayFileNamePart)parameter);
-        if (!isSucceed)
-        {
-            PublishTip(isSucceed);
-            return;
-        }
-
-        SelectedOptionalField = -1;
-    }
-
-    // 可选文件名字段点击事件
-    private RelayCommand<object>? _optionalFieldsCommand;
-
-    public RelayCommand<object> OptionalFieldsCommand => _optionalFieldsCommand ??= RequiredParameterCommand.Create<object>(ExecuteOptionalFieldsCommand);
-
-    /// <summary>
-    /// 可选文件名字段点击事件
-    /// </summary>
-    /// <param name="parameter"></param>
-    private void ExecuteOptionalFieldsCommand(object parameter)
-    {
-        if (SelectedOptionalField == -1)
-        {
-            return;
-        }
-
-        SelectedFileName.Add((DisplayFileNamePart)parameter);
-
-        var fileName = SelectedFileName.Select(item => item.Id).ToImmutableArray();
-        var isSucceed = UpdateVideo(settings => settings with
-        {
-            FileNameParts = fileName
-        }).FileNameParts.SequenceEqual(fileName);
-        PublishTip(isSucceed);
-
-        SelectedOptionalField = -1;
-    }
-
-    // 重置选中文件名字段
-    private RelayCommand? _resetCommand;
-    public RelayCommand ResetCommand => _resetCommand ??= new RelayCommand(ExecuteResetCommand);
-
-    /// <summary>
-    /// 重置选中文件名字段
-    /// </summary>
-    private void ExecuteResetCommand()
-    {
-        var updated = UpdateVideo(settings => settings with
-        {
-            FileNameParts = ApplicationSettingsDefaults.FileNameParts
-        });
-        var isSucceed = updated.FileNameParts.SequenceEqual(ApplicationSettingsDefaults.FileNameParts);
-        PublishTip(isSucceed);
-
-        var fileNameParts = updated.FileNameParts;
-        SelectedFileName.Clear();
-        foreach (var item in fileNameParts)
-        {
-            var display = DisplayFileNamePart(item);
-            SelectedFileName.Add(new DisplayFileNamePart { Id = item, Title = display });
-        }
-
-        SelectedOptionalField = -1;
-    }
-
-    // 文件命名中的时间格式事件
-    private RelayCommand<object>? _fileNamePartTimeFormatCommand;
-
-    public RelayCommand<object> FileNamePartTimeFormatCommand => _fileNamePartTimeFormatCommand ??= RequiredParameterCommand.Create<object>(ExecuteFileNamePartTimeFormatCommand);
-
-    /// <summary>
-    /// 文件命名中的时间格式事件
-    /// </summary>
-    /// <param name="parameter"></param>
-    private void ExecuteFileNamePartTimeFormatCommand(object parameter)
-    {
-        if (parameter is not string timeFormat)
-        {
-            return;
-        }
-
-        var isSucceed = UpdateVideo(settings => settings with
-        {
-            FileNamePartTimeFormat = timeFormat
-        }).FileNamePartTimeFormat == timeFormat;
-        PublishTip(isSucceed);
-    }
-
-    // 文件命名中的序号格式事件
-    private RelayCommand<object>? _orderFormatCommand;
-
-    public RelayCommand<object> OrderFormatCommand => _orderFormatCommand ??= RequiredParameterCommand.Create<object>(ExecuteOrderFormatCommandCommand);
-
-    /// <summary>
-    /// 文件命名中的序号格式事件
-    /// </summary>
-    /// <param name="parameter"></param>
-    private void ExecuteOrderFormatCommandCommand(object parameter)
-    {
-        if (parameter is not OrderFormatDisplay orderFormatDisplay)
-        {
-            return;
-        }
-
-        var isSucceed = UpdateVideo(settings => settings with
-        {
-            OrderFormat = orderFormatDisplay.OrderFormat
-        }).OrderFormat == orderFormatDisplay.OrderFormat;
         PublishTip(isSucceed);
     }
